removed `DistrictRepository.cs` and added logic in `DistrictService.cs` instead
